Move city search to service worker
