+++ src/lib/models.ts
@@ -1,0 +1,8 @@
+export type City = {
+	id: number;
+	name: string;
+	latitude: number;
+	longitude: number;
+	countryCode: string;
+	population: number;
+}
